Require host and port inputs

+++ config/profiling/run-profiling.sh
@@ -1,9 +1,6 @@
 #!/bin/bash
 
-# Take host and port as inputs
-# Take low, medium, high as inputs to set # threads
-# Allow overriding dataset and threads for more info
-# Allow overriding directory name
+# Take low, medium, high as inputs to set # threads - check val and set threads
 
 load="high"
 threads_flag=
@@ -23,6 +20,18 @@ while getopts "l:d:h:p:t:o:" flag; do
         o) dirName=./$OPTARG ;; # use current directory to reduce change of accidentally deleted dir
     esac
 done
+
+if [ -z "$host" ]
+then
+    echo "Host parameter must be set with -h flag"
+    exit 1
+fi
+
+if [ -z "$port" ]
+then
+    echo "Port parameter must be set with -p flag"
+    exit 1
+fi
 
 rm -rf $dirName # jmeter fails if output directory already exists
